requirements.txt
rich>=13.7.1
pyfiglet>=1.0.2
colorama>=0.4.6
requests>=2.32.3
httpx>=0.27.0
alive-progress>=3.1.5
ollama>=0.3.3
psutil>=6.1.0
pyelftools>=0.31
pwntools>=4.13.0

